Implement a webhook to configure cgroup drivers
Depending on the selected SUSE chost version, the cgroup driver
is configured to be cgroupfs for cgroup v1 OS images and systemd
for cgroup v2 OS images.

Cgroup compatibility is determined by the OS image version. For
cHost 15 SP4 and older, it is always cgroup v1, for cHost 15 SP6
it is always cgroup v2. For cHost 15 SP5, anything with a build
timestamp before 20241011 is cgroup v1, anything after is cgroup
v2.